- Release 9.0 - A new major release of Corrade - now with a GUI configurator. :-)
This release adds a GUI configurator to Corrade that allows you to configure all Corrade's settings conveniently - as well as creating an MD5 hash when a password is typed in plaintext. On the other hand, the configuration "Corrade.ini" now comes in a different, albeit similar, format. Nevertheless, the Corrade Configurator will be able to load old configuration files using the "Load Legacy..." button and then save them to the new format using "Save...". This change in style comes from the need to standardize Corrade's configuration as well as provide an easy start without having too much knowledge of Corrade's internals.

Functionally, there are two new commands "setconfigurationdata" and "getconfigurationdata" that are able to set, respectively get Corrade settings using scripts. The interface to changing Corrade's settings no longer requires knowledge of XPATH and all operations are performed using the same semantics as the rest of Corrade's "data"-related commands.